работающая версия без таймера
