+++ js/data/game-data.js
@@ -1,0 +1,8 @@
+const gameData = {
+  screen: `welcome`,
+  questions: 10,
+  timer: 2,
+  lives: 3,
+  score: 0
+};
+export default gameData;
