+++ js/screens/screen--result--fail.js
@@ -1,5 +1,5 @@
 import getElementFromTemplate from '../utils/get-element-from-template';
-import setScreen from '../controllers/set-screen';
+import {onRepeatGame} from '../controllers/game-controller';
 
 export default () => {
   const template = `
@@ -16,7 +16,7 @@ export default () => {
   const replayButton = screenResultFail.querySelector(`.main-replay`);
 
   const onClickReplayButton = function () {
-    setScreen();
+    onRepeatGame();
   };
 
   replayButton.addEventListener(`click`, onClickReplayButton);
