+++ js/utils/get-random-item.js
@@ -1,0 +1,6 @@
+const getRandomItem = function (array) {
+  const randomIndex = Math.floor(Math.random() * array.length);
+  return array[randomIndex];
+};
+
+export default getRandomItem;
